Fix compiler errors caused by internalized methods

+++ core/usecases/src/main/kotlin/storyevent/responses.kt
@@ -5,4 +5,4 @@ import java.util.*
 
 class StoryEventItem(val storyEventId: StoryEvent.Id, val storyEventName: String, val time: Long)
 
-fun StoryEvent.toItem() = StoryEventItem(id, name.value, time)
+fun StoryEvent.toItem() = StoryEventItem(id, name.value, time.toLong())
